Added full text of "Regole generali di traduzione"
This is the first document linked in the current guide (https://www.mozillaitalia.org/home/regole-generali-di-traduzione/)
For the final version we need all the documentation in one place. Will need minor changes.
Note: I omitted the mini-glossary at the end of the document, because it has already been incorporated in gialloporpora's glossary.